usermods: Fix 7sd_reloaded cross module binding
Fix up the cross module binding for
usermods/seven_segment_display_reloaded.  This requires splitting off
headers for BH1750_v2 and SN_Photoresistor.
