+++ usermods/seven_segment_display_reloaded/setup_deps.py
@@ -1,0 +1,9 @@
+Import('env')
+
+
+usermods = env.GetProjectOption("custom_usermods","").split()
+# Check for partner usermods
+if "SN_Photoresistor" in usermods:
+    env.Append(CPPDEFINES=[("USERMOD_SN_PHOTORESISTOR")])
+if "BH1750_v2" in usermods:
+    env.Append(CPPDEFINES=[("USERMOD_BH1750")])
